GPIO interrupts working finally
My goodness, this took a few days to track down and actually understand
the bug that was preventing interrupts from firing.  Actually, there was
nothing wrong and interrupts were working great - the problem was that
the ISR I had made just printed some text to stdout, which is buffered
by default, and I had inadvertantly left off the newline character at
the end of the string. Since the loop I was stalling in never did anything,
the buffer was never actually written to the serial port.  If I had
looked at it in a debugger or done something else, it would have
revealed the fact that my interrupt service routine was in fact being
entered, I was just masking it by accident.