Initial TTC overflow example working
This basically just hooks up the interrupt handler for a TTC and lets it
fly with the defaults (i.e., in overflow mode, no options changed except
wave output).  I suspect that if I had an ILA in this, I'd see a pulse
come out the other side on the wave form generator.  Which makes me
think I should probably add a hardware ILA to this design so I can see
that.