TTC interval finally works - counts 100 sec and matches the wall
The fundamental problem is that the Xilinx SDK does not appropriately
configure the clocks or counter control registers.  Specifically, when
attempting to set interval mode in the control register, it sets the
external clock edge bit in the clock control register instead.  The
code behind the `XTtcPs_SetOptions()` function is Byzantine and makes
absolutely no sense whatsoever to me.  I wrote debug code that is
included to manually poke those registers appropriately and it worked
immediately.  I don't understand why this is so hard for teh Xilinx
folks to get right.