Finaly got a basic TTC-based PWM based software application written
So many problems with this, and I'm still not sure that I don't have a
lingering bug out there.  The SDK TTC driver does not correctly set
options, that much has already been made clear.  Something is still
unsettling about the POR behavior - I have encountered places where
relaunching the same application hangs the application after configuring
the timer.  There isn't a Xilinx provided method for resetting the
timer that I could find in their API, so I just wrote some debug
routines and smashed the register values with the defaults prior to
creating the TTC object that is used by the `XTtcPs` driver. So far, it
seems to have worked, but I haven't found a way to reliably get it to
hang and my reset actions might have been sufficient to solve the
problem.  If anyone ever reads this, and I doubt anyone will, be aware
that the Xilinx hardware timers need to be babysat appropriately, or you
will wind up with odd behavior - this is true for the private timers,
the watchdog, the TTC modules, and presumably the global and SWDT as
well.

If I were using those hardware modules for any sort of mission critical
application, I would not write my own driver and not trust anything that
I got from the vendor.