Set PWM to a 100ms period with a 50% duty cycle - did some other testing with other frequencies and duty cycles too.  Verified everything with an oscilloscope.
Something is still very odd with the interrupt handling following
relaunching which requires a POR to get interrupts to consistently
handled.  I suspect that there is something about not stopping the timer
prior to relaunching the application, but I have no idea.  The problem
is not clearly reproducible, but it's clear that relaunching the
application, particularly after modifying it (even if in a trivial way,
like changing a string that gets printed) and relaunching it causes the
application not service interrupts.  Interestingly, the LED and the
outputs are still oscillating, so there is something going on in the
hardware that only gets cleared or reset by a POR.