Adding working IP source code for the adder / multiplier module that Adam Taylor references in part 24 and so forth of the Microzed Chronicles
This commit contains the following:
- VHDL files based on the machine generated AXI4-Lite peripheral source
code. Includes some basic arithemetic and control sequential logic -
nothing special.
- Some C code that exercises the hardware in the PL to demonstrate some
basic functionality. I'm also including the header file that is
generated by the AXI4-Lite peripheral generator, but I'm not using any
of it.  It's essentially just a couple of syntactic sugar on top of
Xilinx syntactic sugar, which is intended to make it clearer that you
are performing reads and writes into the memory assigned to your
particular peripheral. I guess this is supposed to make life easier
rather than just magical numbers, which is I guess something that gets
abstracted away when you start adding in things like kernels, device
drivers, memory virtualization, and probably a slew of other things I
know nothing about.

Adam has made his source code available in a github repository
somewhere, but it isn't in a one to one correspondence with the material
he covers, so if you're looking at the Microzed Chronicles, this code,
his code, and your code trying to follow along, expect to see some
discrepancies.  Part of the struggle in learning a lot of this stuff is
figuring out how to do what you want with what you're given.