Working example of the altitude to pressure calculations that Taylor makes in Microzed Chronicles sections on PL peripherals.
- The PS/PL example Adam uses computes altitude vs pressure in mbar
using a quadratic equation that is performed in software and then timed
using one of the CPU private timers. He then dumps out some poorly
formatted summary output to show that it takes an average of 25
CPU_3x2x clocks to calculate the result.
- I get something similar, however both of us get a significantly
different result the firt time that the function gets called.  Adam got
89 clocks on his first conversion, and then 25 on the subsequent calls.
- My result is similar, but different because I'm actually using a
function call to do it, so I am assuming that the cycle differences are
due to overhead required to call the function itself.  I'm going to
rewrite it using a macro instead of the function call so that I can see
the difference in overhead.
- My suspicion is that I will see a loss of 5-6 clock cycles due to
removal of the function call overhead.  The difference between the first
and later calls has me stumped though.  I might actually need to step
through the assembly language generated by the compiler to see what's
going on.